EmployeesController Authorization

+++ FoodFusion/Controllers/EmployeesController.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Services.Authorization;
 using Services.Employees;
@@ -10,6 +6,7 @@ using Services.Employees.Exceptions;
 
 namespace WebApi.Controllers
 {
+    [Authorize]
     [Route("api/Restaurants/{restaurantId}")]
     [ApiController]
     public class EmployeesController : ControllerBase
@@ -18,7 +15,7 @@ namespace WebApi.Controllers
         private readonly IEmployeesService _employeesService;
 
         public EmployeesController(
-            IResourceAuthorizationService<EmployeeAuthorizationRequirement> authorizationService, 
+            IResourceAuthorizationService<EmployeeAuthorizationRequirement> authorizationService,
             IEmployeesService employeesService)
         {
             _authorizationService = authorizationService;
@@ -29,6 +26,13 @@ namespace WebApi.Controllers
         [HttpGet("Employees")]
         public IActionResult GetEmployees(int restaurantId)
         {
+            var isAuthorized = _authorizationService
+                .WithUser(User)
+                .WithRequirement(Operations<EmployeeAuthorizationRequirement>.Read)
+                .WithResource(restaurantId)
+                .IsAuthorized();
+            if (!isAuthorized) return Forbid();
+
             var employees = _employeesService.GetEmployees(restaurantId);
             return Ok(employees);
         }
@@ -37,6 +41,13 @@ namespace WebApi.Controllers
         [HttpPost("Employees")]
         public IActionResult PostEmployee(int restaurantId, [FromBody] int userId)
         {
+            var isAuthorized = _authorizationService
+                .WithUser(User)
+                .WithRequirement(Operations<EmployeeAuthorizationRequirement>.Create)
+                .WithResource(restaurantId)
+                .IsAuthorized();
+            if (!isAuthorized) return Forbid();
+
             var employee = _employeesService.AddEmployee(restaurantId, userId);
             return Ok(employee);
         }
@@ -47,6 +58,13 @@ namespace WebApi.Controllers
         {
             try
             {
+                var isAuthorized = _authorizationService
+                    .WithUser(User)
+                    .WithRequirement(Operations<EmployeeAuthorizationRequirement>.Delete)
+                    .WithResource(restaurantId)
+                    .IsAuthorized();
+                if (!isAuthorized) return Forbid();
+
                 _employeesService.RemoveEmployee(restaurantId, userId);
                 return Ok();
             }
@@ -62,6 +80,13 @@ namespace WebApi.Controllers
         {
             try
             {
+                var isAuthorized = _authorizationService
+                    .WithUser(User)
+                    .WithRequirement(Operations<EmployeeAuthorizationRequirement>.Read)
+                    .WithResource(restaurantId)
+                    .IsAuthorized();
+                if (!isAuthorized) return Forbid();
+
                 var manager = _employeesService.GetManager(restaurantId);
                 return Ok(manager);
             }
@@ -75,6 +100,13 @@ namespace WebApi.Controllers
         [HttpPost("Manager")]
         public IActionResult PostManager(int restaurantId, [FromBody] int userId)
         {
+            var isAuthorized = _authorizationService
+                .WithUser(User)
+                .WithRequirement(Operations<EmployeeAuthorizationRequirement>.Create)
+                .WithResource(restaurantId)
+                .IsAuthorized();
+            if (!isAuthorized) return Forbid();
+
             var manager = _employeesService.AddOrReplaceManager(restaurantId, userId);
             return Ok(manager);
         }
